+++ sdk/identity/identity/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Release History
 
-## 1.5.1 (Unreleased)
+## 1.5.1 (2021-08-12)
 
+- Fixed how we verify the IMDS endpoint is available. Now, besides skipping the `Metadata` header, we skip the URL query. Both will ensure that all the known IMDS endpoints return as early as possible.
+- Added support for the `AZURE_POD_IDENTITY_AUTHORITY_HOST` environment variable. If present, the IMDS endpoint initial verification will be skipped.
 
 ## 1.5.0 (2021-07-19)